requirements.txt
# Core
python-dotenv==1.0.1
requests==2.32.3
loguru==0.7.2

# Speech to Text
sounddevice==0.4.7
faster-whisper==1.0.3
numpy==2.1.3

# Text to Speech
pyttsx3==2.98

# Gemini API
google-generativeai==0.8.3

# Optional (for later phases)
duckduckgo-search==6.3.7
customtkinter==5.2.2
opencv-python==4.10.0.84
